FinanceCopilot-Backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.31.0
google-generativeai==0.8.3
faiss-cpu==1.7.4

